[rust-error] Move the unified error into its own crate
