+++ src-rs/oneil/Cargo.toml
@@ -10,6 +10,7 @@ repository.workspace = true
 
 [dependencies]
 oneil_ast = { path = "../oneil_ast" }
+oneil_error = { path = "../oneil_error" }
 oneil_ir = { path = "../oneil_ir" }
 oneil_model_loader = { path = "../oneil_model_loader" }
 oneil_parser = { path = "../oneil_parser" }
