+++ src-rs/oneil_error/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "oneil_error"
+description = "A unified structure for errors in the Oneil programming language"
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
+license.workspace = true
+repository.workspace = true
+
+[dependencies]
